Fix sum method to use int instead of float

+++ methods/addfunctions.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Add2functions {
 	
 public static int sum(int a,int b ) {
-	float sum = a+b;
+	int sum = a+b;
 	return sum;
 }
